Remove Cryptographer - move Cryptographer.encrypt/decrypt to CryptoInfo.encrypted/decrypted.
