+++ src/main/java/org/mozilla/gecko/sync/crypto/Cryptographer.java
@@ -38,141 +38,10 @@
 
 package org.mozilla.gecko.sync.crypto;
 
-import java.security.GeneralSecurityException;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
-import java.util.Arrays;
-
-import javax.crypto.BadPaddingException;
-import javax.crypto.Cipher;
-import javax.crypto.IllegalBlockSizeException;
-import javax.crypto.Mac;
-import javax.crypto.NoSuchPaddingException;
-import javax.crypto.spec.IvParameterSpec;
-import javax.crypto.spec.SecretKeySpec;
-
-import org.mozilla.apache.commons.codec.binary.Base64;
-import org.mozilla.gecko.sync.Utils;
 
 /*
  * Implements the basic required cryptography options.
  */
 public class Cryptographer {
 
-  private static final String TRANSFORMATION     = "AES/CBC/PKCS5Padding";
-  private static final String KEY_ALGORITHM_SPEC = "AES";
-
-  public static CryptoInfo encrypt(CryptoInfo info) throws CryptoException {
-
-    Cipher cipher = getCipher();
-    try {
-      byte[] encryptionKey = info.getKeys().getEncryptionKey();
-      SecretKeySpec spec = new SecretKeySpec(encryptionKey, KEY_ALGORITHM_SPEC);
-
-      // If no IV is provided, we allow the cipher to provide one.
-      if (info.getIV() == null ||
-          info.getIV().length == 0) {
-        cipher.init(Cipher.ENCRYPT_MODE, spec);
-      } else {
-        System.out.println("IV is " + info.getIV().length);
-        cipher.init(Cipher.ENCRYPT_MODE, spec, new IvParameterSpec(info.getIV()));
-      }
-    } catch (GeneralSecurityException ex) {
-      throw new CryptoException(ex);
-    }
-
-    // Encrypt.
-    byte[] encryptedBytes = commonCrypto(cipher, info.getMessage());
-    info.setMessage(encryptedBytes);
-
-    // Save IV.
-    info.setIV(cipher.getIV());
-
-    // Generate HMAC.
-    try {
-      info.setHMAC(info.generatedHMAC());
-    } catch (NoSuchAlgorithmException e) {
-      throw new CryptoException(e);
-    } catch (InvalidKeyException e) {
-      throw new CryptoException(e);
-    }
-
-    return info;
-
-  }
-
-  /*
-   * Perform a decryption.
-   *
-   * @argument info info bundle for decryption
-   *
-   * @return decrypted byte[]
-   *
-   * @throws CryptoException
-   */
-  public static byte[] decrypt(CryptoInfo info) throws CryptoException {
-
-    // Check HMAC.
-    try {
-      if (!info.generatedHMACIsHMAC()) {
-        throw new HMACVerificationException();
-      }
-    } catch (NoSuchAlgorithmException e) {
-      throw new CryptoException(e);
-    } catch (InvalidKeyException e) {
-      throw new CryptoException(e);
-    }
-
-    Cipher cipher = getCipher();
-    try {
-      byte[] encryptionKey = info.getKeys().getEncryptionKey();
-      SecretKeySpec spec = new SecretKeySpec(encryptionKey, KEY_ALGORITHM_SPEC);
-      cipher.init(Cipher.DECRYPT_MODE, spec, new IvParameterSpec(info.getIV()));
-    } catch (GeneralSecurityException ex) {
-      ex.printStackTrace();
-      throw new CryptoException(ex);
-    }
-    return commonCrypto(cipher, info.getMessage());
-  }
-
-  /*
-   * Performs functionality common to both the encryption and decryption
-   * operations.
-   *
-   * Input: Cipher object, non-BaseXX-encoded byte[] input Output:
-   * encrypted/decrypted byte[]
-   */
-  private static byte[] commonCrypto(Cipher cipher, byte[] inputMessage)
-                        throws CryptoException {
-    byte[] outputMessage = null;
-    try {
-      outputMessage = cipher.doFinal(inputMessage);
-    } catch (IllegalBlockSizeException e) {
-      e.printStackTrace();
-      throw new CryptoException(e);
-    } catch (BadPaddingException e) {
-      e.printStackTrace();
-      throw new CryptoException(e);
-    }
-    return outputMessage;
-  }
-
-  /*
-   * Helper to get a Cipher object.
-   * Input: None.
-   * Output: Cipher object.
-   */
-  private static Cipher getCipher() throws CryptoException {
-    Cipher cipher = null;
-    try {
-      cipher = Cipher.getInstance(TRANSFORMATION);
-    } catch (NoSuchAlgorithmException e) {
-      e.printStackTrace();
-      throw new CryptoException(e);
-    } catch (NoSuchPaddingException e) {
-      e.printStackTrace();
-      throw new CryptoException(e);
-    }
-    return cipher;
-  }
 }
